streamlit-ui/requirements.txt
mistralai==0.0.8
streamlit==1.29.0
python-dotenv==1.0.0
together